Add files via upload
This is a python text calculator. In this program the user is able to input numbers and make calculations. There are four different operators that the user can choose from. The user can also decide whether they want to continue using the calculator  or if they would like to finish their session. This program uses multiple types of functions to modulorize the code.  It also uses dictionaries , loops and different types of statements.